Extract analytics helpers as package
